Cross-validation

Confusion matrix

SDeMo.ConfusionMatrix — Type

ConfusionMatrix{T <: Number}

A structure to store the true positives, true negatives, false positives, and false negatives counts (or proportion) during model evaluation. Empty confusion matrices can be created using the zero method.

Folds

These methods will all take as input a vector of labels and a matrix of features, and return a vector of tuples, that have the training indices in the first position, and the validation data in the second. This is not true for holdout, which returns a single tuple.

SDeMo.holdout — Function

holdout(y, X; proportion = 0.2, permute = true)

Sets aside a proportion (given by the proportion keyword, defaults to 0.2) of observations to use for validation, and the rest for training. An additional argument permute (defaults to true) can be used to shuffle the order of observations before they are split.

This method returns a single tuple with the training data first and the validation data second.

holdout(sdm::SDM)

Version of holdout using the instances and labels of an SDM.

SDeMo.montecarlo — Function

montecarlo(y, X; n = 100, kwargs...)

Returns n (def. 100) samples of holdout. Other keyword arguments are passed to holdout.

This method returns a vector of tuples, with each entry have the training data first, and the validation data second.

montecarlo(sdm::SDM)

Version of montecarlo using the instances and labels of an SDM.

SDeMo.leaveoneout — Function

leaveoneout(y, X)

Returns the splits for leave-one-out cross-validation. Each sample is used once, on its own, for validation.

This method returns a vector of tuples, with each entry have the training data first, and the validation data second.

leaveoneout(sdm::SDM)

Version of leaveoneout using the instances and labels of an SDM.

SDeMo.kfold — Function

kfold(y, X; k = 10, permute = true)

Returns splits of the data in which 1 group is used for validation, and k-1 groups are used for training. All kgroups have the (approximate) same size, and each instance is only used once for validation (andk`-1 times for training).

This method returns a vector of tuples, with each entry have the training data first, and the validation data second.

kfold(sdm::SDM)

Version of kfold using the instances and labels of an SDM.

Cross-validation

SDeMo.crossvalidate — Function

crossvalidate(sdm, folds; thr = nothing, kwargs...)

Performs cross-validation on a model, given a vector of tuples representing the data splits. The threshold can be fixed through the thr keyword arguments. All other keywords are passed to the train! method.

This method returns two vectors of ConfusionMatrix, with the confusion matrix for each set of validation data first, and the confusion matrix for the training data second.

Null classifiers

SDeMo.noskill — Function

noskill(labels::Vector{Bool})

Returns the confusion matrix for the no-skill classifier given a vector of labels. Predictions are made at random, with each class being selected by its proportion in the training data.

noskill(sdm::SDM)

Version of noskill using the training labels for an SDM.

SDeMo.coinflip — Function

coinflip(labels::Vector{Bool})

Returns the confusion matrix for the no-skill classifier given a vector of labels. Predictions are made at random, with each class being selected with a probability of one half.

coinflip(sdm::SDM)

Version of coinflip using the training labels for an SDM.

SDeMo.constantnegative — Function

constantnegative(labels::Vector{Bool})

Returns the confusion matrix for the constant positive classifier given a vector of labels. Predictions are assumed to always be negative.

constantnegative(sdm::SDM)

Version of constantnegative using the training labels for an SDM.

SDeMo.constantpositive — Function

constantpositive(labels::Vector{Bool})

Returns the confusion matrix for the constant positive classifier given a vector of labels. Predictions are assumed to always be positive.

constantpositive(sdm::SDM)

Version of constantpositive using the training labels for an SDM.

List of performance measures

SDeMo.tpr — Function

tpr(M::ConfusionMatrix)

True-positive rate

$\frac{TP}{TP+FN}$

SDeMo.tnr — Function

tnr(M::ConfusionMatrix)

True-negative rate

$\frac{TN}{TN+FP}$

SDeMo.fpr — Function

fpr(M::ConfusionMatrix)

False-positive rate

$\frac{FP}{FP+TN}$

SDeMo.fnr — Function

fnr(M::ConfusionMatrix)

False-negative rate

$\frac{FN}{FN+TP}$

SDeMo.ppv — Function

ppv(M::ConfusionMatrix)

Positive predictive value

$\frac{TP}{TP+FP}$

SDeMo.npv — Function

npv(M::ConfusionMatrix)

Negative predictive value

$\frac{TN}{TN+FN}$

SDeMo.fdir — Function

fdir(M::ConfusionMatrix)

False discovery rate, 1 - ppv

SDeMo.fomr — Function

fomr(M::ConfusionMatrix)

False omission rate, 1 - npv

SDeMo.plr — Function

plr(M::ConfusionMatrix)

Positive likelihood ratio

$\frac{TPR}{FPR}$

SDeMo.nlr — Function

nlr(M::ConfusionMatrix)

Negative likelihood ratio

$\frac{FNR}{TNR}$

SDeMo.accuracy — Function

accuracy(M::ConfusionMatrix)

Accuracy

$\frac{TP + TN}{TP + TN + FP + FN}$

SDeMo.balancedaccuracy — Function

balanced(M::ConfusionMatrix)

Balanced accuracy

$\frac{1}{2} (TPR + TNR)$

SDeMo.f1 — Function

f1(M::ConfusionMatrix)

F₁ score, defined as the harmonic mean between precision and recall:

$2\times\frac{PPV\times TPR}{PPV + TPR}$

This uses the more general fscore internally.

SDeMo.fscore — Function

fscore(M::ConfusionMatrix, β=1.0)

Fᵦ score, defined as the harmonic mean between precision and recall, using a positive factor β indicating the relative importance of recall over precision:

$(1 + \beta^2)\times\frac{PPV\times TPR}{(\beta^2 \times PPV) + TPR}$

SDeMo.trueskill — Function

trueskill(M::ConfusionMatrix)

True skill statistic (a.k.a Youden's J, or informedness)

$TPR + TNR - 1$

SDeMo.markedness — Function

markedness(M::ConfusionMatrix)

Markedness, a measure similar to informedness (TSS) that emphasizes negative predictions

$PPV + NPV -1$

SDeMo.dor — Function

dor(M::ConfusionMatrix)

Diagnostic odd ratio, defined as plr/nlr. A useful test has a value larger than unity, and this value has no upper bound.

SDeMo.κ — Function

κ(M::ConfusionMatrix)

Cohen's κ

SDeMo.mcc — Function

mcc(M::ConfusionMatrix)

Matthew's correlation coefficient. This is the default measure of model performance, and there are rarely good reasons to use anything else to decide which model to use.

Confidence interval

SDeMo.ci — Function

ci(C::Vector{ConfusionMatrix}, f)

Applies f to all confusion matrices in the vector, and returns the 95% CI.

ci(C::Vector{ConfusionMatrix})

Applies the MCC (mcc) to all confusion matrices in the vector, and returns the 95% CI.

Aliases

SDeMo.specificity — Function

specificity(M::ConfusionMatrix)

Alias for tnr, the true negative rate

SDeMo.sensitivity — Function

sensitivity(M::ConfusionMatrix)

Alias for tpr, the true positive rate

SDeMo.recall — Function

recall(M::ConfusionMatrix)

Alias for tpr, the true positive rate

SDeMo.precision — Function

precision(M::ConfusionMatrix)

Alias for ppv, the positive predictive value