Ensembles

Bagging

SDeMo.Bagging — Type

Bagging

SDeMo.outofbag — Function

outofbag(ensemble::Bagging; kwargs...)

This method returns the confusion matrix associated to the out of bag error, wherein the succes in predicting instance i is calculated on the basis of all models that have not been trained on i. The consensus of the different models is a simple majority rule.

The additional keywords arguments are passed to predict.

SDeMo.bootstrap — Function

bootstrap(y, X; n = 50)

bootstrap(sdm::SDM; kwargs...)

SDeMo.iqr — Function

iqr(x, m=0.25, M=0.75)

Returns the inter-quantile range, by default between 25% and 75% of observations.

Heterogeneous ensembles

SDeMo.Ensemble — Type

Ensemble

An heterogeneous ensemble model is defined as a vector of SDMs.