Internals

GBIF.format_date — Function

Internal function to format dates in records

GBIF.validate_occurrence_query — Function

validate_occurrence_query(query::Pair)

Checks that the queries for occurrences searches are well formatted.

This is used internally.

Everything this function does is derived from the GBIF API documentation, including (and especially) the values for enum types. This modifies the queryset. Filters that are not allowed are removed, and filters that have incorrect values are dropped too.

This feels like the most conservative option – the user can always filter the results when they are returned.

Base.show — Function

Show an occurrence

show(io::IO, o::GBIFRecord)

Displays the key, the taxon name, and the country of observation.

Show several occurrences

show(io::IO, o::GBIFRecords)

Displays the total number, and the number of currently unmasked records.

Show a taxonomic record

show(io::IO, t::GBIFTaxon)

Displays the taxon name.