Basic information on layers

SimpleSDMLayers.latitudes — Function

latitudes(layer::T) where {T <: SimpleSDMLayer}

Returns an iterator with the latitudes of the SDM layer passed as its argument. This returns the latitude at the center of each cell in the grid.

SimpleSDMLayers.latitudes(record::GBIF.GBIFRecord)

Returns the latitude associated to a GBIF record

SimpleSDMLayers.latitudes(records::GBIF.GBIFRecords)

Returns the non-missing latitudes from a series of records

SimpleSDMLayers.longitudes — Function

longitudes(layer::T) where {T <: SimpleSDMLayer}

Returns an iterator with the longitudes of the SDM layer passed as its argument. This returns the longitudes at the center of each cell in the grid.

SimpleSDMLayers.longitudes(record::GBIF.GBIFRecord)

Returns the longitude associated to a GBIF record

SimpleSDMLayers.latitudes(records::GBIF.GBIFRecords)

Returns the non-missing latitudes from a series of records

SimpleSDMLayers.boundingbox — Function

boundingbox(layer::T) where {T <: SimpleSDMLayer}

Returns the bounding coordinates of a layer as NamedTuple.

SimpleSDMLayers.grid — Function

grid(layer::T) where {T <: SimpleSDMLayer}

Returns a copy of the grid for a given layer.

SimpleSDMLayers.cellsize — Function

cellsize(layer::SimpleSDMLayer)

Returns a layer of the same type (predictor or response) where the value of each cell is the surface area of this cell assuming that the Earth is round (not "round as opposed to flat", but "round as opposed to potato shaped").

In practice, this function works by measuring the area of the spherical cap at the top and bottom of each cell, then subtracting one from the other, to get the area of the ribbon for this range of latitudes. This is divided by the number of cells needed to cover the entire ribbon, which is given by the ratio between the longitudinal size of each cell and 360 (being the breadth of possible longitudes).