Operations on layers

SimpleSDMLayers.clip — Function

clip(layer::T, p1::Point, p2::Point; expand=Symbol[]) where {T <: SimpleSDMLayer}

Return a raster by defining a bounding box through two points. The order of the points (in terms of bottom/top/left/right) is not really important, as the correct coordinates will be extracted.

This method is the one that other versions of clip ultimately end up calling.

This operation takes an additional keyword argument expand, which is a vector of Symbols, the symbols being any combination of :left, :right, :bottom, and :top. This argument specifies, in the case where the coordinate to clip is at the limit between two cells in a raster, whether the outer neighboring row/column should be included. This defaults to expand=Symbol[], i.e. the outer neighboring cell will not be included.

clip(layer::T; left=nothing, right=nothing, top=nothing, bottom=nothing, kwargs...) where {T <: SimpleSDMLayer}

Clips a raster by giving the (optional) limits left, right, bottom, and top.

This operation takes an additional keyword argument expand, which is a vector of Symbols, the symbols being any combination of :left, :right, :bottom, and :top. This argument specifies, in the case where the coordinate to clip is at the limit between two cells in a raster, whether the outer neighboring row/column should be included. This defaults to expand=Symbol[], i.e. the outer neighboring cell will not be included.

clip(origin::T1, destination::T2; kwargs...) where {T1 <: SimpleSDMLayer, T2 <: SimpleSDMLayer}

Clips a layer by another layer, i.e. subsets the first layer so that it has the dimensions of the second layer.

This operation takes an additional keyword argument expand, which is a vector of Symbols, the symbols being any combination of :left, :right, :bottom, and :top. This argument specifies, in the case where the coordinate to clip is at the limit between two cells in a raster, whether the outer neighboring row/column should be included. This defaults to expand=Symbol[], i.e. the outer neighboring cell will not be included.

clip(layer::T, records::GBIF.GBIFRecords)

Returns a clipped version (with a 10% margin) around all occurrences in a GBIFRecords collection.

SimpleSDMLayers.coarsen — Function

coarsen(L::LT, f::FT, d::Tuple{IT,IT}) where {LT <: SimpleSDMLayer, FT <: Function, IT <: Integer}

This function will aggregate a layer by merging cells. The function f passed as its second argument is expected to take an array as input, and return a single value of any type (but it is sort of a social contract here that this will be a number of some sort, and if it is not, that you really know what you're doing).

The size of the cells to aggregate is given by the tuple, so that (2,2) will make groups of two cells vertically and two cells horizontally, for a total of four cells. By default, the cells containing nothing will be ignored, so that f is only applied to non-nothing values.

In cases where the number of cells to aggregate is not matching with the size of the grid, and ArgumentError will be thrown. Note that there is no expectation that the two values in d will be the same.

SimpleSDMLayers.slidingwindow — Function

slidingwindow(L::LT, f::FT, d::IT) where {LT <: SimpleSDMLayer, FT <: Function, IT <: Number}

This function will replace the value at any cell by applying the function f to the array of cells values that are within a distance d (in kilometers) from the focal cell. This is, for example, useful to use an average to smooth out the layers. The distance is estimated using the haversine distance, assuming that the radius of the Earth is 6371.0 km. This means that the size of the window will vary a little bit across latitudes, but this is far better than using a number of cells, which would have dramatic consequences near the poles.

It always returns a SimpleSDMResponse, and the cells containing nothing will also not contain a value in the output. This is different from the behavior of coarsen, which tends to expand the area of the layer in which we have data.

This function is currently relatively slow. Performance improvements will arrive at some point.

SimpleSDMLayers.mask — Function

mask(l1::T1, l2::T2) where {T1 <: SimpleSDMLayer, T2 <: SimpleSDMLayer}

Returns a copy of the second layer in which the positions for which the first layer is zero (of the appropriate type) or nothing are set to nothing. This is mostly useful in cases where you have a Bool layer.

mask(layer::SimpleSDMLayer, records::GBIF.GBIFRecords, element_type::Type=Bool)

Create a new layer storing information about the presence of occurrences in the cells, either counting (numeric types) or presence-absence-ing (boolean types) them.

SimpleSDMLayers.rescale! — Function

rescale!(layer::TI, template::TJ) where {TI <: SimpleSDMLayer, TJ <: SimpleSDMLayer}

Changes the values of the layer given as its first argument, so that it has the same range as the values of the layer given as its second argument. Modification is done in-place.

rescale!(layer::TI, t::Tuple{T,T}) where {TI <: SimpleSDMLayer, T <: Number}

Changes the values of the layer given as its first argument, so that it has the same range as the values given as a tuple of values. Modification is done in-place.

rescale!(layer::T, p::Vector{Real}) where {T <: SimpleSDMLayer}

Rescale the values of a layer so that they match with the quantiles given in p. Internally, this uses the Statistics.quantile function.

SimpleSDMLayers.rescale — Function

rescale(layer::TI, template::TJ) where {TI <: SimpleSDMLayer, TJ <: SimpleSDMLayer}

Copying version of rescale!.

rescale(layer::TI, t::Tuple{T,T}) where {TI <: SimpleSDMLayer, T <: Number}

Copying version of rescale!.

rescale(layer::T, p::Vector{Real}) where {T <: SimpleSDMLayer}

Copying version of rescale!.

SimpleSDMLayers.mosaic — Function

mosaic(f::TF, layers::Vector{T}) where {TF <: Function, T <: SimpleSDMLayer}

Joins a series of possibly overlapping layers by applying the function f to the values that occupy the same cells. Note that the function f should return a single value and accept a vector as an input. Functions like Statistics.mean, etc, work well.

Using mosaic with maximum is equivalent to raster::merge from the R package raster.

SimpleSDMLayers.tile — Function

tile(layer::T, s::Tuple{Integer, Integer} = (5, 5)) where {T <: SimpleSDMLayer}

Split a layer into a matrix of tiles (with size s), where the tiles are as evenly sized as possible, and have strictly matching boundaries. This function will allocate the return matrix.

SimpleSDMLayers.tile! — Function

tile!(tiles::Matrix{T}, layer::T, s::Tuple{Integer, Integer} = (5, 5)) where {T <: SimpleSDMLayer}

Split a layer into a matrix of tiles (with size s), where the tiles are as evenly sized as possible, and have strictly matching boundaries.

SimpleSDMLayers.stitch — Function

stitch(tiles::Matrix{T})

Returns a layer from a series of layer tiles, as produced by tile or tile!.