List of types

SimpleSDMLayers.SimpleSDMLayer — Type

All types in the package are part of the abstract type SimpleSDMLayer. A SimpleSDMLayer has five core fields: grid is a matrix storing the cells, and left, right, bottom and top are floating point numbers specifying the bounding box.

It is assumed that the missing values will be represented as nothing, so internally the matrix will have type Union{T, Nothing}.

SimpleSDMLayers.SimpleSDMResponse — Type

A response is a SimpleSDMLayer that is mutable, and is the usual type to store analysis outputs. You can transform a response into a predictor using convert.

SimpleSDMLayers.SimpleSDMPredictor — Type

A predictor is a SimpleSDMLayer that is immutable, and so does not have methods for setindex!, etc. It is a safe way to store values that should not be modified by the analysis. Note that if you are in a bind, the values of the grid field are not immutable, but don't tell anyone we told you. The correct way of handling predictors you need to modify would be to use convert methods.

Missing docstring.

Missing docstring for RasterCell. Check Documenter's build log for details.