Additional operations of GBIF data

SimpleSDMLayers.longitudes — Method

SimpleSDMLayers.latitudes(records::GBIF.GBIFRecords)

Returns the non-missing latitudes from a series of records

SimpleSDMLayers.latitudes — Method

SimpleSDMLayers.latitudes(records::GBIF.GBIFRecords)

Returns the non-missing latitudes from a series of records

Missing docstring.

Missing docstring for SimpleSDMLayers.mask. Check Documenter's build log for details.

SimpleSDMLayers.mask! — Function

mask!(l1::T1, l2::T2) where {T1 <: SimpleSDMLayer, T2 <: SimpleSDMLayer}

Changes the second layer so that the positions for which the first layer is false (for Boolean layers) or nothing are set to nothing. This is mostly useful in cases where you have a Bool layer.

mask!(layer::SimpleSDMResponse{T}, records::GBIF.GBIFRecords) where {T <: AbstractBool}

Fills a layer (most likely created with similar) so that the values are true if an occurrence is found in the cell, false if not.

mask!(layer::SimpleSDMResponse{T}, records::GBIF.GBIFRecords) where {T <: Number}

Fills a layer (most likely created with similar) so that the values reflect the number of occurrences in the cell.