Pseudo absences generation

Supported algorithms

SpeciesDistributionToolkit.PseudoAbsenceGenerator — Type

PseudoAbsenceGenerator

Abstract type to which all of the pseudo-absences generator types belong. Note that the pseudo-absence types are singleton types, and the arguments are passed when generating the pseudo-absence mask.

SpeciesDistributionToolkit.WithinRadius — Type

WithinRadius

Generates pseudo-absences within a set radius (in kilometers) around each occurrence

SpeciesDistributionToolkit.SurfaceRangeEnvelope — Type

SurfaceRangeEnvelope

Generates pseudo-absences at random within the geographic range covered by actual occurrences

SpeciesDistributionToolkit.RandomSelection — Type

RandomSelection

Generates pseudo-absences at random within the layer

Generation of a pseudo-absence mask

The above algorithms are used in conjunction with pseudoabsencemask to generate a Boolean layer that contains all pixels in which a background point can be. They do not generate background points directly, in order to allow more flexible workflows based on clipping Boolean masks, for example.

SpeciesDistributionToolkit.pseudoabsencemask — Function

pseudoabsencemask(::Type{RandomSelection}, presence::T) where {T <: SDMLayer}

Generates a mask for pseudo-absences using the random selection method. Candidate cells for the pseudo-absence mask are (i) within the bounding box of the layer (use SurfaceRangeEnvelope to use the presences bounding box), and (ii) valued in the layer.

pseudoabsencemask( ::Type{SurfaceRangeEnvelope}, presences::T) where {T <: SDMLayer}

Generates a mask from which pseudo-absences can be drawn, by picking cells that are (i) within the bounding box of occurrences, (ii) valued in the layer, and (iii) not already occupied by an occurrence

pseudoabsencemask(::Type{DistanceToEvent}, presence::T; f=minimum) where {T <: SimpleSDMLayer}

Generates a mask for pseudo-absences using the distance to event method. Candidate cells are weighted according to their distance to a known observation, with far away places being more likely. Depending on the distribution of distances, it may be a very good idea to flatten this layer using log or an exponent. The f function is used to determine which distance is reported (minimum by default, can also be mean or median).

Sampling of background points

SpeciesDistributionToolkit.backgroundpoints — Function

backgroundpoints(layer::T, n::Int; replace::Bool=false, kwargs...) where {T <: SimpleSDMLayer}

Generates background points based on a layer that gives either the location of possible points (Bool) or the weight of each cell in the final sample (Number). Note that the default value is to draw without replacement, but this can be changed using replace=true. The additional keywors arguments are passed to StatsBase.sample, which is used internally.