Transformers and classifiers

Transformers (univariate)

SDeMo.RawData — Type

RawData

A transformer that does nothing to the data. This is passing the raw data to the classifier.

SDeMo.ZScore — Type

ZScore

A transformer that scales and centers the data, using only the data that are avaiable to the model at training time. For all variables in the SDM features (regardless of whether they are used), this transformer will store the observed mean and standard deviation.

Transformers (multivariate)

SDeMo.MultivariateTransform — Type

MultivariateTransform{T} <: Transformer

T is a multivariate transformation, likely offered through the MultivariateStats package. The transformations currently supported are PCA, PPCA, KernelPCA, and Whitening.

Classifiers

SDeMo.NaiveBayes — Type

NaiveBayes

Naive Bayes Classifier

By default, upon training, the prior probability will be set to the prevalence of the training data.

SDeMo.BIOCLIM — Type

BIOCLIM

BIOCLIM

SDeMo.DecisionTree — Type

DecisionTree

TODO