Overloads

Base.:- — Function

Base.-(l1::SimpleSDMLayer, l2::SimpleSDMLayer)

Applies - (from Base) to every pair of cells from two SimpleSDMLayers and returns the result as a new SimpleSDMResponse layer. Note that - is only applied to the pairs without a nothing element, and returns nothing for the pairs with one. This function has been automatically generated.

Base.:* — Function

Base.*(l1::SimpleSDMLayer, l2::SimpleSDMLayer)

Applies * (from Base) to every pair of cells from two SimpleSDMLayers and returns the result as a new SimpleSDMResponse layer. Note that * is only applied to the pairs without a nothing element, and returns nothing for the pairs with one. This function has been automatically generated.

Base.:/ — Function

Base./(l1::SimpleSDMLayer, l2::SimpleSDMLayer)

Applies / (from Base) to every pair of cells from two SimpleSDMLayers and returns the result as a new SimpleSDMResponse layer. Note that / is only applied to the pairs without a nothing element, and returns nothing for the pairs with one. This function has been automatically generated.

Base.:+ — Function

Base.+(l1::SimpleSDMLayer, l2::SimpleSDMLayer)

Applies + (from Base) to every pair of cells from two SimpleSDMLayers and returns the result as a new SimpleSDMResponse layer. Note that + is only applied to the pairs without a nothing element, and returns nothing for the pairs with one. This function has been automatically generated.

Missing docstring.

Missing docstring for `Base.:==

Base.convert — Function

Base.convert(::Type{SimpleSDMResponse}, layer::T) where {T <: SimpleSDMPredictor}

Returns a response with the same grid and bounding box as the predictor.

Base.convert(::Type{SimpleSDMPredictor}, layer::T) where {T <: SimpleSDMResponse}

Returns a predictor with the same grid and bounding box as the response.

Base.convert(::Type{T}, layer::TL) where {T <: Number, TL <: SimpleSDMLayer}

Returns a copy of the layer with the same type (response or predictor), but the element type has been changed to T (which must be a number). This function is extremely useful (required, in fact) for plotting, as the nothing values are changed to NaN in the heatmaps.

Base.convert(::Type{Matrix}, layer::T) where {T <: SimpleSDMLayer}

Returns the grid as an array.

Base.copy — Function

Base.copy(l::T) where {T <: SimpleSDMLayer}

Returns a new copy of the layer, which has the same type.

Base.eachindex — Function

Base.eachindex(layer::T) where {T <: SimpleSDMLayer}

Returns the index of the grid.

Base.eltype — Function

Base.eltype(layer::SimpleSDMLayer{T}) where {T}

Returns the type of the RasterCell describing each entry in the layer.

Base.exp — Function

Base.exp(l::SimpleSDMResponse{T}) where {T <: Number}

Applies exp (from Base) to every cell within a SimpleSDMResponse, as long as this cell is not nothing. This function has been automatically generated.

Base.exp(l::SimpleSDMPredictor{T}) where {T <: Number}

Applies exp (from Base) to every cell within a SimpleSDMPredictor, as long as this cell is not nothing. This function has been automatically generated.

Base.exp10 — Function

Base.exp10(l::SimpleSDMResponse{T}) where {T <: Number}

Applies exp10 (from Base) to every cell within a SimpleSDMResponse, as long as this cell is not nothing. This function has been automatically generated.

Base.exp10(l::SimpleSDMPredictor{T}) where {T <: Number}

Applies exp10 (from Base) to every cell within a SimpleSDMPredictor, as long as this cell is not nothing. This function has been automatically generated.

Base.exp2 — Function

Base.exp2(l::SimpleSDMResponse{T}) where {T <: Number}

Applies exp2 (from Base) to every cell within a SimpleSDMResponse, as long as this cell is not nothing. This function has been automatically generated.

Base.exp2(l::SimpleSDMPredictor{T}) where {T <: Number}

Applies exp2 (from Base) to every cell within a SimpleSDMPredictor, as long as this cell is not nothing. This function has been automatically generated.

Base.expm1 — Function

Base.expm1(l::SimpleSDMResponse{T}) where {T <: Number}

Applies expm1 (from Base) to every cell within a SimpleSDMResponse, as long as this cell is not nothing. This function has been automatically generated.

Base.expm1(l::SimpleSDMPredictor{T}) where {T <: Number}

Applies expm1 (from Base) to every cell within a SimpleSDMPredictor, as long as this cell is not nothing. This function has been automatically generated.

Base.extrema — Function

Base.extrema(l::SimpleSDMResponse{T}) where {T <: Number}

Applies extrema (from Base) to an object of type SimpleSDMResponse. This function has been automatically generated. Note that this function is only applied to the non-nothing elements of the layer, and has no method to work on the dims keyword; the grid itself can be extracted with convert(Matrix, l).

Base.extrema(l::SimpleSDMPredictor{T}) where {T <: Number}

Applies extrema (from Base) to an object of type SimpleSDMPredictor. This function has been automatically generated. Note that this function is only applied to the non-nothing elements of the layer, and has no method to work on the dims keyword; the grid itself can be extracted with convert(Matrix, l).

Base.getindex — Function

Base.getindex(layer::T, i::CartesianIndex{2}) where {T <: SimpleSDMLayer}

Returns the value stored at a given cartesian index.

Base.getindex(layer::T, i::Integer) where {T <: SimpleSDMLayer}

Returns the value stored at a linear index (a good candidate for deprecation as we have a better iteration interface now).

Base.getindex(layer::T, i::Integer, j::Integer) where {T <: SimpleSDMLayer}

Standard abstract array accession, where the dimensions follow the dimensions of the underlying grid.

Base.getindex(layer::T, i::Array{CartesianIndex{2}}) where {T <: SimpleSDMLayer}

Returns an array of values based on an array of cartesian indices - this can be a vector or a matrix, and the elements can be in any order. This will not return a raster.

Base.getindex(layer::T, longitude::AbstractFloat, latitude::AbstractFloat) where {T <: SimpleSDMLayer}

Returns a value by longitude and latitude.

Base.getindex(layer::T, c::Point) where {T <: SimpleSDMLayer}

Access a value through a Point (from GeometryBasics), which has the longitude first and the latitude last – this follows the GeoJSON convention.

Base.getindex(layer::T, c::Array{<:Point}) where {T <: SimpleSDMLayer}

Access a value through an array of Point (from GeometryBasics), which has the longitude first and the latitude last. The array can be in any order, so this method will not return a raster.

Base.getindex(::T, ::Nothing) where {T <: SimpleSDMLayer}

If the user requests a point that is out of bounds, its cartesian coordinate will be matched to nothing, and then we return nothing.

Base.hcat — Function

Base.hcat(l1::T, l2::T) where {T <: SimpleSDMLayers}

Adds the second layer to the right of the first one (according to coordinates), assuming the strides and left/right coordinates match. This will automatically re-order the layers if the second is to the left the first.

Base.isequal — Function

isequal(layer1::SimpleSDMLayer, layer2::SimpleSDMLayer)

Tests whether two SimpleSDMLayer elements are equal. The layers are equal if all their fields (grid, left, right, bottom, top) are equal, as verified with isequal (e.g., isequal(layer1.grid, layer2.grid)).

Base.IteratorsMD.CartesianIndices — Type

Base.CartesianIndices(layer::T) where {T <: SimpleSDMLayer}

This function is used to access the positions in a layer as cartesian indices. Internally, although it may be more convenient to access positions by their coordinates, the value of the raster are extracted using their cartesian coordinates.

Base.keys — Function

Base.keys(layer::T) where {T <: SimpleSDMLayer}

Returns an array of Point where every entry in the array is a non-nothing grid coordinate.

Base.log — Function

Base.log(l::SimpleSDMResponse{T}) where {T <: Number}

Applies log (from Base) to every cell within a SimpleSDMResponse, as long as this cell is not nothing. This function has been automatically generated.

Base.log(l::SimpleSDMPredictor{T}) where {T <: Number}

Applies log (from Base) to every cell within a SimpleSDMPredictor, as long as this cell is not nothing. This function has been automatically generated.

Base.log10 — Function

Base.log10(l::SimpleSDMResponse{T}) where {T <: Number}

Applies log10 (from Base) to every cell within a SimpleSDMResponse, as long as this cell is not nothing. This function has been automatically generated.

Base.log10(l::SimpleSDMPredictor{T}) where {T <: Number}

Applies log10 (from Base) to every cell within a SimpleSDMPredictor, as long as this cell is not nothing. This function has been automatically generated.

Base.log1p — Function

Base.log1p(l::SimpleSDMResponse{T}) where {T <: Number}

Applies log1p (from Base) to every cell within a SimpleSDMResponse, as long as this cell is not nothing. This function has been automatically generated.

Base.log1p(l::SimpleSDMPredictor{T}) where {T <: Number}

Applies log1p (from Base) to every cell within a SimpleSDMPredictor, as long as this cell is not nothing. This function has been automatically generated.

Base.log2 — Function

Base.log2(l::SimpleSDMResponse{T}) where {T <: Number}

Applies log2 (from Base) to every cell within a SimpleSDMResponse, as long as this cell is not nothing. This function has been automatically generated.

Base.log2(l::SimpleSDMPredictor{T}) where {T <: Number}

Applies log2 (from Base) to every cell within a SimpleSDMPredictor, as long as this cell is not nothing. This function has been automatically generated.

Base.max — Function

Base.max(l1::SimpleSDMLayer, l2::SimpleSDMLayer)

Applies max (from Base) to every pair of cells from two SimpleSDMLayers and returns the result as a new SimpleSDMResponse layer. Note that max is only applied to the pairs without a nothing element, and returns nothing for the pairs with one. This function has been automatically generated.

Base.maximum — Function

Base.maximum(l::SimpleSDMResponse{T}) where {T <: Number}

Applies maximum (from Base) to an object of type SimpleSDMResponse. This function has been automatically generated. Note that this function is only applied to the non-nothing elements of the layer, and has no method to work on the dims keyword; the grid itself can be extracted with convert(Matrix, l).

Base.maximum(l::SimpleSDMPredictor{T}) where {T <: Number}

Applies maximum (from Base) to an object of type SimpleSDMPredictor. This function has been automatically generated. Note that this function is only applied to the non-nothing elements of the layer, and has no method to work on the dims keyword; the grid itself can be extracted with convert(Matrix, l).

Base.min — Function

Base.min(l1::SimpleSDMLayer, l2::SimpleSDMLayer)

Applies min (from Base) to every pair of cells from two SimpleSDMLayers and returns the result as a new SimpleSDMResponse layer. Note that min is only applied to the pairs without a nothing element, and returns nothing for the pairs with one. This function has been automatically generated.

Base.minimum — Function

Base.minimum(l::SimpleSDMResponse{T}) where {T <: Number}

Applies minimum (from Base) to an object of type SimpleSDMResponse. This function has been automatically generated. Note that this function is only applied to the non-nothing elements of the layer, and has no method to work on the dims keyword; the grid itself can be extracted with convert(Matrix, l).

Base.minimum(l::SimpleSDMPredictor{T}) where {T <: Number}

Applies minimum (from Base) to an object of type SimpleSDMPredictor. This function has been automatically generated. Note that this function is only applied to the non-nothing elements of the layer, and has no method to work on the dims keyword; the grid itself can be extracted with convert(Matrix, l).

Base.replace — Function

Base.replace(layer::T, old_new::Pair...) where {T <: SimpleSDMResponse}

Replaces the elements of layer according to a series of pairs. Returns a copy.

Base.replace(layer::T, old_new::Pair...) where {T <: SimpleSDMPredictor}

Replaces the elements of layer according to a series of pairs. Copies the layer as a response before.

Base.replace! — Function

Base.replace!(layer::T, old_new::Pair...) where {T <: SimpleSDMLayer}

Replaces the elements of layer according to a series of pairs. In place. Only possible for SimpleSDMResponse elements (which are mutable) and will throw an error if called on a SimpleSDMPredictor element (which is not mutable).

Base.show — Function

Base.show(io::IO, ::MIME"text/plain", layer::T) where {T <: SimpleSDMLayer}
Base.show(io::IO, layer::T) where {T <: SimpleSDMLayer}

Shows a textual representation of the layer.

Base.similar — Function

Base.similar(layer::T, ::Type{TC}) where {TC <: Any, T <: SimpleSDMLayer}

Returns a SimpleSDMResponse of the same dimensions as the original layer, with nothing in the same positions. The rest of the values are replaced by the output of zero(TC), which implies that there must be a way to get a zero for the type. If not, the same result can always be achieved through the use of copy, manual update, and convert.

Base.similar(layer::T) where {T <: SimpleSDMLayer}

Returns a SimpleSDMResponse of the same dimensions as the original layer, with nothing in the same positions. The rest of the values are replaced by the output of zero(element_type), which implies that there must be a way to get a zero for the type. If not, the same result can always be achieved through the use of copy, manual update, and convert.

Base.size — Function

Base.size(layer::T) where {T <: SimpleSDMLayer}

Returns the size of the grid.

Base.size(layer::T, i...) where {T <: SimpleSDMLayer}

Returns the size of the grid alongside a dimension.

Base.sqrt — Function

Base.sqrt(l::SimpleSDMResponse{T}) where {T <: Number}

Applies sqrt (from Base) to every cell within a SimpleSDMResponse, as long as this cell is not nothing. This function has been automatically generated.

Base.sqrt(l::SimpleSDMPredictor{T}) where {T <: Number}

Applies sqrt (from Base) to every cell within a SimpleSDMPredictor, as long as this cell is not nothing. This function has been automatically generated.

Base.stride — Function

Base.stride(layer::T; dims::Union{Nothing,Integer}=nothing) where {T <: SimpleSDMLayer}

Returns the stride, i.e. half the length, of cell dimensions, possibly alongside a side of the grid. The first position is the length of the longitude cells, the second the latitude.

Base.sum — Function

Base.sum(l::SimpleSDMResponse{T}) where {T <: Number}

Applies sum (from Base) to an object of type SimpleSDMResponse. This function has been automatically generated. Note that this function is only applied to the non-nothing elements of the layer, and has no method to work on the dims keyword; the grid itself can be extracted with convert(Matrix, l).

Base.sum(l::SimpleSDMPredictor{T}) where {T <: Number}

Applies sum (from Base) to an object of type SimpleSDMPredictor. This function has been automatically generated. Note that this function is only applied to the non-nothing elements of the layer, and has no method to work on the dims keyword; the grid itself can be extracted with convert(Matrix, l).

Base.values — Function

Base.values(l::T) where {T <: SimpleSDMLayer}

Returns the non-nothing values of a layer.

Base.vcat — Function

Base.vcat(l1::T, l2::T) where {T <: SimpleSDMLayers}

Adds the second layer under the first one (according to coordinates), assuming the strides and left/right coordinates match. This will automatically re-order the layers if the second is above the first.

SimpleSDMLayers._layers_are_compatible — Function

_layers_are_compatible(l1::X, l2::Y) where {X <: SimpleSDMLayer, Y <: SimpleSDMLayer}

Internal function to verify if layers are compatible, i.e. have the same
size and bounding coordinates.

_layers_are_compatible(layers::Array{T}) where {T <: SimpleSDMLayer}

Are the layers compatibles?

SimpleSDMLayers.mask! — Function

mask!(l1::T1, l2::T2) where {T1 <: SimpleSDMLayer, T2 <: SimpleSDMLayer}

Changes the second layer so that the positions for which the first layer is false (for Boolean layers) or nothing are set to nothing. This is mostly useful in cases where you have a Bool layer.

Statistics.mean — Function

mean(itr)

Compute the mean of all elements in a collection.

Note

If itr contains NaN or missing values, the result is also NaN or missing (missing takes precedence if array contains both). Use the skipmissing function to omit missing entries and compute the mean of non-missing values.

Examples

julia> using Statistics

julia> mean(1:20)
10.5

julia> mean([1, missing, 3])
missing

julia> mean(skipmissing([1, missing, 3]))
2.0

mean(f::Function, itr)

Apply the function f to each element of collection itr and take the mean.

julia> using Statistics

julia> mean(√, [1, 2, 3])
1.3820881233139908

julia> mean([√1, √2, √3])
1.3820881233139908

mean(f::Function, A::AbstractArray; dims)

Apply the function f to each element of array A and take the mean over dimensions dims.

Julia 1.3

This method requires at least Julia 1.3.

julia> using Statistics

julia> mean(√, [1, 2, 3])
1.3820881233139908

julia> mean([√1, √2, √3])
1.3820881233139908

julia> mean(√, [1 2 3; 4 5 6], dims=2)
2×1 Matrix{Float64}:
 1.3820881233139908
 2.2285192400943226

mean(A::AbstractArray; dims)

Compute the mean of an array over the given dimensions.

Julia 1.1

mean for empty arrays requires at least Julia 1.1.

Examples

julia> using Statistics

julia> A = [1 2; 3 4]
2×2 Matrix{Int64}:
 1  2
 3  4

julia> mean(A, dims=1)
1×2 Matrix{Float64}:
 2.0  3.0

julia> mean(A, dims=2)
2×1 Matrix{Float64}:
 1.5
 3.5

Statistics.mean(l::SimpleSDMResponse{T}) where {T <: Number}

Applies mean (from Statistics) to an object of type SimpleSDMResponse. This function has been automatically generated. Note that this function is only applied to the non-nothing elements of the layer, and has no method to work on the dims keyword; the grid itself can be extracted with convert(Matrix, l).

Statistics.mean(l::SimpleSDMPredictor{T}) where {T <: Number}

Applies mean (from Statistics) to an object of type SimpleSDMPredictor. This function has been automatically generated. Note that this function is only applied to the non-nothing elements of the layer, and has no method to work on the dims keyword; the grid itself can be extracted with convert(Matrix, l).

Statistics.mean(layers::Array{T}) where {T <: SimpleSDMLayer}

Applies mean (from Statistics) to the elements in corresponding positions from the different layers (similar to mean(a::Array{Matrix})) and returns the result as a new SimpleSDMResponse layer. Note that mean is only applied to the positions without a nothing element and returns nothing for the pairs with one. This function has been automatically generated.

Statistics.median — Function

median(itr)

Compute the median of all elements in a collection. For an even number of elements no exact median element exists, so the result is equivalent to calculating mean of two median elements.

Note

If itr contains NaN or missing values, the result is also NaN or missing (missing takes precedence if itr contains both). Use the skipmissing function to omit missing entries and compute the median of non-missing values.

Examples

julia> using Statistics

julia> median([1, 2, 3])
2.0

julia> median([1, 2, 3, 4])
2.5

julia> median([1, 2, missing, 4])
missing

julia> median(skipmissing([1, 2, missing, 4]))
2.0

median(A::AbstractArray; dims)

Compute the median of an array along the given dimensions.

Examples

julia> using Statistics

julia> median([1 2; 3 4], dims=1)
1×2 Matrix{Float64}:
 2.0  3.0

Statistics.median(l::SimpleSDMResponse{T}) where {T <: Number}

Applies median (from Statistics) to an object of type SimpleSDMResponse. This function has been automatically generated. Note that this function is only applied to the non-nothing elements of the layer, and has no method to work on the dims keyword; the grid itself can be extracted with convert(Matrix, l).

Statistics.median(l::SimpleSDMPredictor{T}) where {T <: Number}

Applies median (from Statistics) to an object of type SimpleSDMPredictor. This function has been automatically generated. Note that this function is only applied to the non-nothing elements of the layer, and has no method to work on the dims keyword; the grid itself can be extracted with convert(Matrix, l).

Statistics.quantile — Function

quantile(itr, p; sorted=false, alpha::Real=1.0, beta::Real=alpha)

Compute the quantile(s) of a collection itr at a specified probability or vector or tuple of probabilities p on the interval [0,1]. The keyword argument sorted indicates whether itr can be assumed to be sorted.

Samples quantile are defined by Q(p) = (1-γ)*x[j] + γ*x[j+1], where $x[j]$ is the j-th order statistic, and γ is a function of j = floor(n*p + m), m = alpha + p*(1 - alpha - beta) and g = n*p + m - j.

By default (alpha = beta = 1), quantiles are computed via linear interpolation between the points ((k-1)/(n-1), v[k]), for k = 1:n where n = length(itr). This corresponds to Definition 7 of Hyndman and Fan (1996), and is the same as the R and NumPy default.

The keyword arguments alpha and beta correspond to the same parameters in Hyndman and Fan, setting them to different values allows to calculate quantiles with any of the methods 4-9 defined in this paper:

Def. 4: alpha=0, beta=1

Def. 5: alpha=0.5, beta=0.5

Def. 6: alpha=0, beta=0 (Excel PERCENTILE.EXC, Python default, Stata altdef)

Def. 7: alpha=1, beta=1 (Julia, R and NumPy default, Excel PERCENTILE and PERCENTILE.INC, Python 'inclusive')

Def. 8: alpha=1/3, beta=1/3

Def. 9: alpha=3/8, beta=3/8

Note

An ArgumentError is thrown if v contains NaN or missing values. Use the skipmissing function to omit missing entries and compute the quantiles of non-missing values.

References

Hyndman, R.J and Fan, Y. (1996) "Sample Quantiles in Statistical Packages", The American Statistician, Vol. 50, No. 4, pp. 361-365

Quantile on Wikipedia details the different quantile definitions

Examples

julia> using Statistics

julia> quantile(0:20, 0.5)
10.0

julia> quantile(0:20, [0.1, 0.5, 0.9])
3-element Vector{Float64}:
  2.0
 10.0
 18.000000000000004

julia> quantile(skipmissing([1, 10, missing]), 0.5)
5.5

quantile(layer::T, p) where {T <: SimpleSDMLayer}

Returns the quantiles of layer at p, using Statistics.quantile.

Statistics.std — Function

std(itr; corrected::Bool=true, mean=nothing[, dims])

Compute the sample standard deviation of collection itr.

The algorithm returns an estimator of the generative distribution's standard deviation under the assumption that each entry of itr is a sample drawn from the same unknown distribution, with the samples uncorrelated. For arrays, this computation is equivalent to calculating sqrt(sum((itr .- mean(itr)).^2) / (length(itr) - 1)). If corrected is true, then the sum is scaled with n-1, whereas the sum is scaled with n if corrected is false with n the number of elements in itr.

If itr is an AbstractArray, dims can be provided to compute the standard deviation over dimensions, and means may contain means for each dimension of itr.

A pre-computed mean may be provided. When dims is specified, mean must be an array with the same shape as mean(itr, dims=dims) (additional trailing singleton dimensions are allowed).

Note

If array contains NaN or missing values, the result is also NaN or missing (missing takes precedence if array contains both). Use the skipmissing function to omit missing entries and compute the standard deviation of non-missing values.

Statistics.std(l::SimpleSDMResponse{T}) where {T <: Number}

Applies std (from Statistics) to an object of type SimpleSDMResponse. This function has been automatically generated. Note that this function is only applied to the non-nothing elements of the layer, and has no method to work on the dims keyword; the grid itself can be extracted with convert(Matrix, l).

Statistics.std(l::SimpleSDMPredictor{T}) where {T <: Number}

Applies std (from Statistics) to an object of type SimpleSDMPredictor. This function has been automatically generated. Note that this function is only applied to the non-nothing elements of the layer, and has no method to work on the dims keyword; the grid itself can be extracted with convert(Matrix, l).

Statistics.std(layers::Array{T}) where {T <: SimpleSDMLayer}

Applies std (from Statistics) to the elements in corresponding positions from the different layers (similar to mean(a::Array{Matrix})) and returns the result as a new SimpleSDMResponse layer. Note that std is only applied to the positions without a nothing element and returns nothing for the pairs with one. This function has been automatically generated.