Type system for datasets

Note that all datasets must be a direct subtype of RasterDataset, all providers must be a direct subtype of RasterProvider, all scenarios must be a direct subtype of FutureScenario, and all models must be a direct subtype of FutureModel. If you need to aggregate multiple models within a type (e.g. CMIP6Scenarios), use a Union type. The reason for this convention is that in interactive mode, subtype will let users pick the data combination they want.

Type system overview

SimpleSDMDatasets.RasterData — Type

RasterData{P <: RasterProvider, D <: RasterDataset}

The RasterData type is the main user-facing type for SimpleSDMDatasets. Specifically, this is a singleton parametric type, where the two parameters are the type of the RasterProvider and the RasterDataset. Note that the inner constructor calls the provides method on the provider/dataset pair to check that this combination exists.

SimpleSDMDatasets.RasterDataset — Type

RasterDataset

This is an abstract type to label something as being a dataset. Datasets are given by RasterProviders, and the same dataset can have multiple providers.

SimpleSDMDatasets.RasterProvider — Type

RasterProvider

This is an abstract type to label something as a provider of RasterDatasets. For example, WorldClim2 and CHELSA2 are RasterProviders.

List of datasets

SimpleSDMDatasets.BioClim — Type

BioClim

SimpleSDMDatasets.Elevation — Type

Elevation

SimpleSDMDatasets.MinimumTemperature — Type

MinimumTemperature

SimpleSDMDatasets.MaximumTemperature — Type

MaximumTemperature

SimpleSDMDatasets.AverageTemperature — Type

AverageTemperature

SimpleSDMDatasets.Precipitation — Type

Precipitation

SimpleSDMDatasets.SolarRadiation — Type

SolarRadiation

SimpleSDMDatasets.WindSpeed — Type

WindSpeed

SimpleSDMDatasets.WaterVaporPressure — Type

WaterVaporPressure

SimpleSDMDatasets.LandCover — Type

LandCover

SimpleSDMDatasets.HabitatHeterogeneity — Type

HabitatHeterogeneity

SimpleSDMDatasets.Topography — Type

Topography

SimpleSDMDatasets.BirdRichness — Type

BirdRichness

SimpleSDMDatasets.MammalRichness — Type

MammalRichness

SimpleSDMDatasets.AmphibianRichness — Type

AmphibianRichness

List of providers

SimpleSDMDatasets.WorldClim2 — Type

WorldClim2

This provider offers access to the version 2 of the WorldClim data, accessible from http://www.worldclim.com/version2.

SimpleSDMDatasets.EarthEnv — Type

EarthEnv

SimpleSDMDatasets.CHELSA1 — Type

CHELSA1

SimpleSDMDatasets.CHELSA2 — Type

CHELSA2

SimpleSDMDatasets.BiodiversityMapping — Type

BiodiversityMapping

Global biodiveristy data from https://biodiversitymapping.org/ - see this website for citation information

SimpleSDMDatasets.PaleoClim — Type

PaleoClim

Paleoclimate data from http://www.paleoclim.org/ - see this website for citation information

List of projections

SimpleSDMDatasets.Projection — Type

Future{S <: FutureScenario, M <: FutureModel}

This type is similar to RasterData but describes a combination of a scenario and a model. Note that unlike RasterData, there is no type check in the inner constructor; instead, the way to check that a provider/dataset/scenario/model combination exists is to overload the provides method for a dataset and future.

SimpleSDMDatasets.CMIP5Scenario — Type

CMIP5Scenario

These scenarios are part of CMIP5. They can be RCP26 to RCP85 (with RCPXX the scenario).

List of enumerated types

SimpleSDMDatasets.RasterDownloadType — Type

RasterDownloadType

This enum stores the possible types of downloaded files. They are listed with instances(RasterDownloadType), and are currently limited to _file (a file, can be read directly) and _zip (an archive, must be unzipped).

SimpleSDMDatasets.RasterFileType — Type

RasterFileType

This enum stores the possible types of returned files. They are listed with instances(RasterFileType).

SimpleSDMDatasets.RasterCRS — Type

RasterCRS

This enum stores the possible coordinate representation system of returned files. They are listed with instances(RasterProjection).