Tools for SDM demos and education

The prediction pipeline

SDeMo.AbstractSDM — Type

AbstractSDM

TODO

SDeMo.AbstractEnsembleSDM — Type

AbstractEnsembleSDM

TODO

SDeMo.SDM — Type

SDM

This type specifies a full model, which is composed of a transformer (which applies a transformation on the data), a classifier (which returns a quantitative score), a threshold (above which the score corresponds to the prediction of a presence).

In addition, the SDM carries with it the training features and labels, as well as a vector of indices indicating which variables are actually used by the model.

SDeMo.Transformer — Type

Transformer

TODO

SDeMo.Classifier — Type

Classifier

TODO

Utility functions

SDeMo.features — Function

features(sdm::SDM)

Returns the features stored in the field X of the SDM. Note that the features are an array, and this does not return a copy of it – any change made to the output of this function will change the content of the SDM features.

features(sdm::SDM, n)

Returns the n-th feature stored in the field X of the SDM.

SDeMo.labels — Function

labels(sdm::SDM)

Returns the labels stored in the field y of the SDM – note that this is not a copy of the labels, but the object itself.

SDeMo.threshold — Function

threshold(sdm::SDM)

This returns the value above which the score returned by the SDM is considered to be a presence.

SDeMo.threshold! — Function

threshold!(sdm::SDM, τ)

Sets the value of the threshold.

SDeMo.variables — Function

variables(sdm::SDM)

Returns the list of variables used by the SDM – these may be ordered by importance. This does not return a copy of the variables array, but the array itself.

SDeMo.variables! — Function

variables!(sdm::SDM, v)

Sets the list of variables.

SDeMo.instance — Function

instance(sdm::SDM, n)

Returns the n-th instance stored in the field X of the SDM.

Training and predicting

SDeMo.train! — Function

train!(ensemble::Bagging; kwargs...)

Trains all the model in an ensemble model - the keyword arguments are passed to train! for each model. Note that this retrains the entire model, which includes the transformers.

train!(ensemble::Ensemble; kwargs...)

Trains all the model in an heterogeneous ensemble model - the keyword arguments are passed to train! for each model. Note that this retrains the entire model, which includes the transformers.

The keywod arguments are passed to train! and can include the training indices.

train!(sdm::SDM; threshold=true, training=:, optimality=mcc)

This is the main training function to train a SDM.

The three keyword arguments are:

training: defaults to :, and is the range (or alternatively the indices) of the data that are used to train the model

threshold: defaults to true, and performs moving threshold by evaluating 200 possible values between the minimum and maximum output of the model, and returning the one that is optimal

optimality: defaults to mcc, and is the function applied to the confusion matrix to evaluate which value of the threshold is the best

Internally, this function trains the transformer, then projects the data, then trains the classifier. If threshold is true, the threshold is then optimized.

StatsAPI.predict — Function

StatsAPI.predict(ensemble::Bagging, X; consensus = median, kwargs...)

Returns the prediction for the ensemble of models a dataset X. The function used to aggregate the outputs from different models is consensus (defaults to median). All other keyword arguments are passed to predict.

To get a direct estimate of the variability, the consensus function can be changed to iqr (inter-quantile range), or any measure of variance.

StatsAPI.predict(ensemble::Bagging; kwargs...)

Predicts the ensemble model for all training data.

StatsAPI.predict(ensemble::Ensemble, X; consensus = median, kwargs...)

Returns the prediction for the heterogeneous ensemble of models a dataset X. The function used to aggregate the outputs from different models is consensus (defaults to median). All other keyword arguments are passed to predict.

To get a direct estimate of the variability, the consensus function can be changed to iqr (inter-quantile range), or any measure of variance.

StatsAPI.predict(ensemble::Ensemble; kwargs...)

Predicts the heterogeneous ensemble model for all training data.

StatsAPI.predict(sdm::SDM, X; threshold = true)

This is the main prediction function, and it takes as input an SDM and a matrix of features. The only keyword argument is threshold, which determines whether the prediction is returned raw or as a binary value (default is true).

StatsAPI.predict(sdm::SDM; kwargs...)

This method performs the prediction on the entire set of training data available for the training of an SDM.

SDeMo.reset! — Function

reset!(sdm::SDM, thr=0.5)

Resets a model, with a potentially specified value of the threshold. This amounts to re-using all the variables, and removing the tuned threshold version.