An interface for occurrence data

Types that other packages should use

OccurrencesInterface.AbstractOccurrence — Type

AbstractOccurrence

Other types describing a single observation can be sub-types.

OccurrencesInterface.AbstractOccurrenceCollection — Type

AbstractOccurrenceCollection

Other types describing multiple observations can be sub-types.

Concrete types shipping with the package

OccurrencesInterface.Occurrence — Type

Occurence

OccurrencesInterface.Occurrences — Type

Occurrences

The interface

OccurrencesInterface.elements — Function

elements(::T) where {T<:AbstractOccurrenceCollection}

Returns the elements contained in an abstract collection of occurrences – this must be something that can be iterated.

OccurrencesInterface.entity — Function

entity(o::Occurrence)

Returns the entity (species name) for an occurrence event.

entity(::AbstractOccurrence)

Default method for any abstract occurrence type for the entity operation. Unless overloaded, this returns nothing.

entity(::AbstractOccurrenceCollection)

Default method for any abstract occurrence collection type for the entity operation. Unless overloaded, this returns an array of entity on all elements of the argument.

OccurrencesInterface.place — Function

place(o::Occurrence)

Returns the place of the occurrence event, either as a tuple of float in the longitude, latitude format, or as missing. The CRS is assumed to the WGS84.

place(::AbstractOccurrence)

Default method for any abstract occurrence type for the place operation. Unless overloaded, this returns nothing.

place(::AbstractOccurrenceCollection)

Default method for any abstract occurrence collection type for the place operation. Unless overloaded, this returns an array of place on all elements of the argument.

OccurrencesInterface.date — Function

date(o::Occurrence)

Returns the date (technically a DateTime object) documenting the time of occurrence event. Can be missing if not known.

date(::AbstractOccurrence)

Default method for any abstract occurrence type for the date operation. Unless overloaded, this returns nothing.

date(::AbstractOccurrenceCollection)

Default method for any abstract occurrence collection type for the date operation. Unless overloaded, this returns an array of date on all elements of the argument.

OccurrencesInterface.presence — Function

presence(o::Occurrence)

Returns a Bool for the occurrence status, where true is the presence of the entity and false is the (pseudo)absence.

presence(::AbstractOccurrence)

Default method for any abstract occurrence type for the presence operation. Unless overloaded, this returns nothing.

presence(::AbstractOccurrenceCollection)

Default method for any abstract occurrence collection type for the presence operation. Unless overloaded, this returns an array of presence on all elements of the argument.

OccurrencesInterface.presences — Function

presences(c::T) where {T<:AbstractOccurrenceCollection}

Returns an Occurrences where only the occurrences in the initial collection for which presence evaluates to true are kept.

OccurrencesInterface.absences — Function

absences(c::T) where {T<:AbstractOccurrenceCollection}

Returns an Occurrences where only the occurrences in the initial collection for which presence evaluates to false are kept.