SimpleSDMPolygons

SimpleSDMDatasets.downloader — Method

downloader

SimpleSDMDatasets.keychecker — Method

keychecker

SimpleSDMDatasets.provides — Method

provides

SimpleSDMDatasets.resolutions — Method

resolutions

SimpleSDMPolygons._extra_keychecks — Method

_extra_keychecks

Used for checking combinations of keyword arguments.

For example, for some Countries providers, different countrys have different available levels, and these provider implement this method to check if the requested (country, level) pair is provided.

SimpleSDMPolygons.destination — Method

destination

SimpleSDMPolygons.levels — Method

levels

SimpleSDMPolygons.postprocess — Method

postprocess

SimpleSDMPolygons.AbstractGeometry — Type

AbstractGeometry

SimpleSDMPolygons.Bioregions — Type

Bioregions

This is a largest unit than Ecoregions, used by OneEarth

SimpleSDMPolygons.Countries — Type

Countries

To be used when retrieving country borders, although GADM uses it for internal admin areas.

SimpleSDMPolygons.EPA — Type

EPA

This data is all hosted by the US Environmental Protection Agency (EPA), so there's really no guarantee it will stay available day-to-day.

SimpleSDMPolygons.ESRI — Type

ESRI

Datasets from the ESRI dataset hub - https://hub.arcgis.com/datasets/.

SimpleSDMPolygons.Ecoregions — Type

Ecoregions

Boundaries that are defined primarily w.r.t. biotic or environmental characteristics, and that usually span across different administrative boundaries.

SimpleSDMPolygons.Feature — Type

Feature

SimpleSDMPolygons.FeatureCollection — Type

FeatureCollection

SimpleSDMPolygons.GADM — Type

GADM

Known for their server's reliable uptime!

SimpleSDMPolygons.Lakes — Type

Lakes

To be used when retrieving the boundaries of lakes and other similar water bodies.

SimpleSDMPolygons.Land — Type

Land

To be used when retrieving land masses, without reference to administrative areas.

SimpleSDMPolygons.MultiPolygon — Type

MultiPolygon

SimpleSDMPolygons.NaturalEarth — Type

NaturalEarth

It's literally a buncha GeoJSONs in a Github repo

SimpleSDMPolygons.Oceans — Type

Oceans

Like Lakes but not like lakes; lakes with more water and salt.

SimpleSDMPolygons.OneEarth — Type

OneEarth

Ecoregions from https://www.oneearth.org/ - note that this is described as "bioregions" by the provider, but is functionally similar to ecoregions. Both BioRegions and EcoRegions work as types.

SimpleSDMPolygons.OpenStreetMap — Type

OpenStreetMap

This provider uses the nominatim OSM Germany service to map a user-provided name of any arbitrary place to an OSM ID, then uses the polygon creation service hosted by OSM France to return a GeoJSON file.

SimpleSDMPolygons.ParksAndProtected — Type

ParksAndProtected

Used for parks and protected areas, regardless of the jurisdiction that establishes them (municipal, regional, national, etc).

SimpleSDMPolygons.Places — Type

Places

To be used when retrieving an arbitrary sized place. This can correspond to very local (smaller admin areas) places, arbitrary locations as with the OSM service, or any other data source where a more focused dataset type does not apply.

SimpleSDMPolygons.Polygon — Type

Polygon

SimpleSDMPolygons.PolygonData — Type

PolygonData

SimpleSDMPolygons.PolygonDataset — Type

PolygonDataset

This is an abstract type for all datasets supported by the package.

SimpleSDMPolygons.PolygonDownloadType — Type

PolygonDownloadType

SimpleSDMPolygons.PolygonFileType — Type

PolygonFileType

SimpleSDMPolygons.PolygonProvider — Type

PolygonProvider

SimpleSDMPolygons.Resolve — Type

Resolve

SimpleSDMPolygons._GeoJSON — Type

_GeoJSON

SimpleSDMPolygons._Shapefile — Type

_Shapefile