OccurrencesInterface

OccurrencesInterface.__demodata — Method

OccurrencesInterface.__demodata()

This dataset has 2150 observations of bigfoot from the Bigfoot Field Researchers Organization, a community science project on cryptids.

OccurrencesInterface.absences — Method

absences(c::T) where {T<:AbstractOccurrenceCollection}

Returns an Occurrences where only the occurrences in the initial collection for which presence evaluates to false are kept.

OccurrencesInterface.date — Method

date(o::Occurrence)

Returns the date (technically a DateTime object) documenting the time of occurrence event. Can be missing if not known.

OccurrencesInterface.date — Method

date(::AbstractOccurrence)

Default method for any abstract occurrence type for the date operation. Unless overloaded, this returns nothing.

OccurrencesInterface.date — Method

date(::AbstractOccurrenceCollection)

Default method for any abstract occurrence collection type for the date operation. Unless overloaded, this returns an array of date on all elements of the argument.

OccurrencesInterface.date — Method

date(::Vector{AbstractOccurrence})

Default method for any vector of occurrence type for the date operation. Unless overloaded, this returns an array of date on all elements of the argument.

OccurrencesInterface.elements — Method

elements(::T) where {T<:AbstractOccurrenceCollection}

Returns the elements contained in an abstract collection of occurrences – this must be something that can be iterated. The default value, when unimplemented, is nothing.

OccurrencesInterface.entity — Method

entity(o::Occurrence)

Returns the entity (species name) for an occurrence event.

OccurrencesInterface.entity — Method

entity(::AbstractOccurrence)

Default method for any abstract occurrence type for the entity operation. Unless overloaded, this returns nothing.

OccurrencesInterface.entity — Method

entity(::AbstractOccurrenceCollection)

Default method for any abstract occurrence collection type for the entity operation. Unless overloaded, this returns an array of entity on all elements of the argument.

OccurrencesInterface.entity — Method

entity(::Vector{AbstractOccurrence})

Default method for any vector of occurrence type for the entity operation. Unless overloaded, this returns an array of entity on all elements of the argument.

OccurrencesInterface.place — Method

place(o::Occurrence)

Returns the place of the occurrence event, either as a tuple of float in the longitude, latitude format, or as missing. The CRS is assumed to be WGS84 with no option to change it. This follows the GeoJSON specification.

OccurrencesInterface.place — Method

place(::AbstractOccurrence)

Default method for any abstract occurrence type for the place operation. Unless overloaded, this returns nothing.

OccurrencesInterface.place — Method

place(::AbstractOccurrenceCollection)

Default method for any abstract occurrence collection type for the place operation. Unless overloaded, this returns an array of place on all elements of the argument.

OccurrencesInterface.place — Method

place(::Vector{AbstractOccurrence})

Default method for any vector of occurrence type for the place operation. Unless overloaded, this returns an array of place on all elements of the argument.

OccurrencesInterface.presence — Method

presence(o::Occurrence)

Returns a Bool for the occurrence status, where true is the presence of the entity and false is the (pseudo)absence.

OccurrencesInterface.presence — Method

presence(::AbstractOccurrence)

Default method for any abstract occurrence type for the presence operation. Unless overloaded, this returns nothing.

OccurrencesInterface.presence — Method

presence(::AbstractOccurrenceCollection)

Default method for any abstract occurrence collection type for the presence operation. Unless overloaded, this returns an array of presence on all elements of the argument.

OccurrencesInterface.presence — Method

presence(::Vector{AbstractOccurrence})

Default method for any vector of occurrence type for the presence operation. Unless overloaded, this returns an array of presence on all elements of the argument.

OccurrencesInterface.presences — Method

presences(c::T) where {T<:AbstractOccurrenceCollection}

Returns an Occurrences where only the occurrences in the initial collection for which presence evaluates to true are kept.

OccurrencesInterface.AbstractOccurrence — Type

AbstractOccurrence

Other types describing a single observation should be sub-types of this. Occurrences are always defined as a single observation of a single species.

OccurrencesInterface.AbstractOccurrenceCollection — Type

AbstractOccurrenceCollection

Other types describing multiple observations can be sub-types of this. Occurrences collections are a way to collect multiple observations of arbitrarily many species.

OccurrencesInterface.Occurrence — Type

Occurrence

This is a sub-type of AbstractOccurrence, with the following types:

what - species name, defaults to ""

presence - a boolean to mark the presence of the species, defaults to true

where - a tuple giving the location as longitude,latitude in WGS84, or missing, defaults to missing

when - a DateTime giving the date of observation, or missing, defaults to missing

When the interface is properly implemented for any type that is a sub-type of AbstractOccurrence, there is an Occurrence object can be created directly with e.g. Occurrence(observation). There is, similarly, an automatically implemented convert method.

OccurrencesInterface.Occurrence — Method

Occurrence(t::T) where {T<:AbstractOccurrence}

Given a type that is a subtype of AbstractOccurrence, constructs an Occurrence object using the methods in the interface. This should not be overloaded in other packages.

OccurrencesInterface.Occurrences — Type

Occurrences

This is a sub-type of AbstractOccurrenceCollection. No default value.