SimpleSDMPolygons

SimpleSDMDatasets.downloader — Method

downloader

SimpleSDMDatasets.keychecker — Method

keychecker

SimpleSDMDatasets.provides — Method

provides

SimpleSDMDatasets.resolutions — Method

resolutions

SimpleSDMPolygons._extra_keychecks — Method

_extra_keychecks

Used for checking combinations of keyword arguments.

For example, for some Countries providers, different countrys have different available levels, and these provider implement this method to check if the requested (country, level) pair is provided.

SimpleSDMPolygons.destination — Method

destination

SimpleSDMPolygons.levels — Method

levels

SimpleSDMPolygons.postprocess — Method

postprocess

SimpleSDMPolygons.AbstractGeometry — Type

AbstractGeometry

SimpleSDMPolygons.Bioregions — Type

Bioregions

This is a largest unit than Ecoregions, used by OneEarth

SimpleSDMPolygons.Countries — Type

Countries

SimpleSDMPolygons.EPA — Type

EPA

This data is all hosted by the US Environmental Protection Agency (EPA), so there's really no guarantee it will stay available day-to-day.

SimpleSDMPolygons.Ecoregions — Type

Ecoregions

SimpleSDMPolygons.Feature — Type

Feature

SimpleSDMPolygons.FeatureCollection — Type

FeatureCollection

SimpleSDMPolygons.GADM — Type

GADM

Known for their server's reliable uptime!

SimpleSDMPolygons.Lakes — Type

Lakes

SimpleSDMPolygons.Land — Type

Land

SimpleSDMPolygons.MultiPolygon — Type

MultiPolygon

SimpleSDMPolygons.NaturalEarth — Type

NaturalEarth

It's literally a buncha GeoJSONs in a Github repo

SimpleSDMPolygons.Oceans — Type

Oceans

SimpleSDMPolygons.OneEarth — Type

OneEarth

Ecoregions from https://www.oneearth.org/ - note that this is described as "bioregions" by the provider, but is functionally similar to ecoregions. Both BioRegions and EcoRegions work as types.

SimpleSDMPolygons.OpenStreetMap — Type

OpenStreetMap

It's specifically openstreetmap.fr, and a python script that does...something...and spits out a GeoJSON. OuvertRueCarte. Is that anything.

SimpleSDMPolygons.ParksAndProtected — Type

ParksAndProtected

SimpleSDMPolygons.Places — Type

Places

SimpleSDMPolygons.Polygon — Type

Polygon

SimpleSDMPolygons.PolygonData — Type

PolygonData

SimpleSDMPolygons.PolygonDataset — Type

PolygonDataset

SimpleSDMPolygons.PolygonDownloadType — Type

PolygonDownloadType

SimpleSDMPolygons.PolygonFileType — Type

PolygonFileType

SimpleSDMPolygons.PolygonProvider — Type

PolygonProvider

SimpleSDMPolygons.Resolve — Type

Resolve

SimpleSDMPolygons._GeoJSON — Type

_GeoJSON

SimpleSDMPolygons._Shapefile — Type

_Shapefile