Internals

Phylopic._get_uuids_at_page — Function

_get_uuids_at_page(query, page)

This function is an internal helped function to return an array of pairs, wherein each pair maps a name to a UUID, for a given query and page. These outpurs are collected in a dictionary by Phylopic.names.