Pseudo absences generation

Supported algorithms

SpeciesDistributionToolkit.PseudoAbsenceGenerator — Type

PseudoAbsenceGenerator

Abstract type to which all of the pseudo-absences generator types belong. Note that the pseudo-absence types are singleton types, and the arguments are passed when generating the pseudo-absence mask.

SpeciesDistributionToolkit.WithinRadius — Type

WithinRadius

Generates pseudo-absences within a set radius (in kilometers) around each occurrence

SpeciesDistributionToolkit.SurfaceRangeEnvelope — Type

SurfaceRangeEnvelope

Generates pseudo-absences at random within the geographic range covered by actual occurrences

SpeciesDistributionToolkit.RandomSelection — Type

RandomSelection

Generates pseudo-absences at random within the layer

Generation of a pseudo-absence mask

The above algorithms are used in conjunction with pseudoabsencemask to generate a Boolean layer that contains all pixels in which a background point can be. They do not generate background points directly, in order to allow more flexible workflows based on clipping Boolean masks, for example.

SpeciesDistributionToolkit.pseudoabsencemask — Function

pseudoabsencemask(::Type{RandomSelection}, presence::T) where {T <: SimpleSDMLayer}

Generates a mask for pseudo-absences using the random selection method. Candidate cells for the pseudo-absence mask are (i) within the bounding box of the layer (use SurfaceRangeEnvelope to use the presences bounding box), and (ii) valued in the layer.

pseudoabsencemask( ::Type{SurfaceRangeEnvelope}, presences::T) where {T <: SimpleSDMLayer}

Generates a mask from which pseudo-absences can be drawn, by picking cells that are (i) within the bounding box of occurrences, (ii) valued in the layer, and (iii) not already occupied by an occurrence

Sampling of background points

SpeciesDistributionToolkit.sample — Function

sample(layer::T, n::Integer = 1; kwargs...) where {T <: SimpleSDMLayer}

Sample a series of background points from a Boolean layer. The kwargs arguments are passed to StatsBase.sample. This method returns a Boolean layer where the values of true correspond to a background point.

sample( layer::T, weights::T2, n::Integer = 1; kwargs..., ) where {T <: SimpleSDMLayer, T2 <: SimpleSDMLayer}

Sample a series of background points from a Boolean layer, where each point has a probability of being included in the background given by the second layer. This methods works like (and is, in fact, a wrapper around) StatsBase.sample, where the cell values in the weight layers are transformed into weights.