Overloads

Missing docstring.

Missing docstring for Base.:-. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.:*. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.:/. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.:+. Check Documenter's build log for details.

Missing docstring.

Missing docstring for `Base.:==

Missing docstring.

Missing docstring for Base.convert. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.copy. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.eachindex. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.eltype. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.exp. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.exp10. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.exp2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.expm1. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.extrema. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.getindex. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.hcat. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.isequal. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.IteratorsMD.CartesianIndices. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.keys. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.log. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.log10. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.log1p. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.log2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.max. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.maximum. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.min. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.minimum. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.replace. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.replace!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.show. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.similar. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.size. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.sqrt. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.stride. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.sum. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.values. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.vcat. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SimpleSDMLayers._layers_are_compatible. Check Documenter's build log for details.

SimpleSDMLayers.mask! — Function

mask!(layer::SDMLayer, template::SDMLayer)

Updates the positions in the first layer to be those that have a value in the second layer.

Statistics.mean — Function

mean(itr)

Compute the mean of all elements in a collection.

Note

If itr contains NaN or missing values, the result is also NaN or missing (missing takes precedence if array contains both). Use the skipmissing function to omit missing entries and compute the mean of non-missing values.

Examples

julia> using Statistics

julia> mean(1:20)
10.5

julia> mean([1, missing, 3])
missing

julia> mean(skipmissing([1, missing, 3]))
2.0

mean(f::Function, itr)

Apply the function f to each element of collection itr and take the mean.

julia> using Statistics

julia> mean(√, [1, 2, 3])
1.3820881233139908

julia> mean([√1, √2, √3])
1.3820881233139908

mean(f::Function, A::AbstractArray; dims)

Apply the function f to each element of array A and take the mean over dimensions dims.

Julia 1.3

This method requires at least Julia 1.3.

julia> using Statistics

julia> mean(√, [1, 2, 3])
1.3820881233139908

julia> mean([√1, √2, √3])
1.3820881233139908

julia> mean(√, [1 2 3; 4 5 6], dims=2)
2×1 Matrix{Float64}:
 1.3820881233139908
 2.2285192400943226

mean(A::AbstractArray; dims)

Compute the mean of an array over the given dimensions.

Julia 1.1

mean for empty arrays requires at least Julia 1.1.

Examples

julia> using Statistics

julia> A = [1 2; 3 4]
2×2 Matrix{Int64}:
 1  2
 3  4

julia> mean(A, dims=1)
1×2 Matrix{Float64}:
 2.0  3.0

julia> mean(A, dims=2)
2×1 Matrix{Float64}:
 1.5
 3.5

Statistics.median — Function

median(itr)

Compute the median of all elements in a collection. For an even number of elements no exact median element exists, so the result is equivalent to calculating mean of two median elements.

Note

If itr contains NaN or missing values, the result is also NaN or missing (missing takes precedence if itr contains both). Use the skipmissing function to omit missing entries and compute the median of non-missing values.

Examples

julia> using Statistics

julia> median([1, 2, 3])
2.0

julia> median([1, 2, 3, 4])
2.5

julia> median([1, 2, missing, 4])
missing

julia> median(skipmissing([1, 2, missing, 4]))
2.0

median(A::AbstractArray; dims)

Compute the median of an array along the given dimensions.

Examples

julia> using Statistics

julia> median([1 2; 3 4], dims=1)
1×2 Matrix{Float64}:
 2.0  3.0

Statistics.quantile — Function

quantile(itr, p; sorted=false, alpha::Real=1.0, beta::Real=alpha)

Compute the quantile(s) of a collection itr at a specified probability or vector or tuple of probabilities p on the interval [0,1]. The keyword argument sorted indicates whether itr can be assumed to be sorted.

Samples quantile are defined by Q(p) = (1-γ)*x[j] + γ*x[j+1], where $x[j]$ is the j-th order statistic, and γ is a function of j = floor(n*p + m), m = alpha + p*(1 - alpha - beta) and g = n*p + m - j.

By default (alpha = beta = 1), quantiles are computed via linear interpolation between the points ((k-1)/(n-1), v[k]), for k = 1:n where n = length(itr). This corresponds to Definition 7 of Hyndman and Fan (1996), and is the same as the R and NumPy default.

The keyword arguments alpha and beta correspond to the same parameters in Hyndman and Fan, setting them to different values allows to calculate quantiles with any of the methods 4-9 defined in this paper:

Def. 4: alpha=0, beta=1

Def. 5: alpha=0.5, beta=0.5

Def. 6: alpha=0, beta=0 (Excel PERCENTILE.EXC, Python default, Stata altdef)

Def. 7: alpha=1, beta=1 (Julia, R and NumPy default, Excel PERCENTILE and PERCENTILE.INC, Python 'inclusive')

Def. 8: alpha=1/3, beta=1/3

Def. 9: alpha=3/8, beta=3/8

Note

An ArgumentError is thrown if v contains NaN or missing values. Use the skipmissing function to omit missing entries and compute the quantiles of non-missing values.

References

Hyndman, R.J and Fan, Y. (1996) "Sample Quantiles in Statistical Packages", The American Statistician, Vol. 50, No. 4, pp. 361-365

Quantile on Wikipedia details the different quantile definitions

Examples

julia> using Statistics

julia> quantile(0:20, 0.5)
10.0

julia> quantile(0:20, [0.1, 0.5, 0.9])
3-element Vector{Float64}:
  2.0
 10.0
 18.000000000000004

julia> quantile(skipmissing([1, 10, missing]), 0.5)
5.5

Statistics.std — Function

std(itr; corrected::Bool=true, mean=nothing[, dims])

Compute the sample standard deviation of collection itr.

The algorithm returns an estimator of the generative distribution's standard deviation under the assumption that each entry of itr is a sample drawn from the same unknown distribution, with the samples uncorrelated. For arrays, this computation is equivalent to calculating sqrt(sum((itr .- mean(itr)).^2) / (length(itr) - 1)). If corrected is true, then the sum is scaled with n-1, whereas the sum is scaled with n if corrected is false with n the number of elements in itr.

If itr is an AbstractArray, dims can be provided to compute the standard deviation over dimensions, and means may contain means for each dimension of itr.

A pre-computed mean may be provided. When dims is specified, mean must be an array with the same shape as mean(itr, dims=dims) (additional trailing singleton dimensions are allowed).

Note

If array contains NaN or missing values, the result is also NaN or missing (missing takes precedence if array contains both). Use the skipmissing function to omit missing entries and compute the standard deviation of non-missing values.