Type system for datasets

Note that all datasets must be a direct subtype of RasterDataset, all providers must be a direct subtype of RasterProvider, all scenarios must be a direct subtype of FutureScenario, and all models must be a direct subtype of FutureModel. If you need to aggregate multiple models within a type (e.g. CMIP6Scenarios), use a Union type. The reason for this convention is that in interactive mode, subtype will let users pick the data combination they want.

Type system overview

SimpleSDMDatasets.RasterData — Type

RasterData{P <: RasterProvider, D <: RasterDataset}

The RasterData type is the main user-facing type for SimpleSDMDatasets. Specifically, this is a singleton parametric type, where the two parameters are the type of the RasterProvider and the RasterDataset. Note that the inner constructor calls the provides method on the provider/dataset pair to check that this combination exists.

SimpleSDMDatasets.RasterDataset — Type

RasterDataset

This is an abstract type to label something as being a dataset. Datasets are given by RasterProviders, and the same dataset can have multiple providers.

SimpleSDMDatasets.RasterProvider — Type

RasterProvider

This is an abstract type to label something as a provider of RasterDatasets. For example, WorldClim2 and CHELSA2 are RasterProviders.

List of datasets

Missing docstring.

Missing docstring for BioClim. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elevation. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MinimumTemperature. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaximumTemperature. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AverageTemperature. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Precipitation. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SolarRadiation. Check Documenter's build log for details.

Missing docstring.

Missing docstring for WindSpeed. Check Documenter's build log for details.

Missing docstring.

Missing docstring for WaterVaporPressure. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LandCover. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HabitatHeterogeneity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Topography. Check Documenter's build log for details.

List of providers

Missing docstring.

Missing docstring for WorldClim2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for EarthEnv. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CHELSA1. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CHELSA2. Check Documenter's build log for details.

List of enumerated types

Missing docstring.

Missing docstring for SimpleSDMDatasets.RasterDownloadType. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SimpleSDMDatasets.RasterFileType. Check Documenter's build log for details.